pyproject.toml
[project]
name = "cell-vacuole-tracking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"cellpose",
	"laptrack",
	"pillow",
	"scikit-image",
	"tifffile",
	"numpy",
	"pandas",
	"tqdm",
	"matplotlib",
	"scipy",
	"networkx",
	"cython",
	"opencv-python",
]
